MVC support, bot structure small refactoring, database moved to server, gRPC removed, server entity controllers added, website development started

+++ Bot/Callbacks/SetDefaultPhotoCallback.cs
@@ -2,6 +2,7 @@
 using BlueBellDolls.Bot.Interfaces;
 using BlueBellDolls.Bot.Settings;
 using BlueBellDolls.Bot.Types;
+using BlueBellDolls.Common.Enums;
 using BlueBellDolls.Common.Interfaces;
 using BlueBellDolls.Common.Models;
 using Microsoft.Extensions.Options;
@@ -13,7 +14,7 @@ namespace BlueBellDolls.Bot.Callbacks
         private readonly IMessagesHelperService _messagesHelperService;
         private readonly IArgumentParseHelperService _argumentParseHelperService;
         private readonly IMessagesProvider _messagesProvider;
-        private readonly IManagementService _managementService;
+        private readonly IManagementServicesFactory _managementServicesFactory;
 
         public SetDefaultPhotoCallback(
             IBotService botService,
@@ -21,38 +22,45 @@ namespace BlueBellDolls.Bot.Callbacks
             ICallbackDataProvider callbackDataProvider,
             IMessagesHelperService messagesHelperService,
             IArgumentParseHelperService argumentParseHelperService,
-            IMessagesProvider messagesProvider,
-            IManagementService managementService)
+            IManagementServicesFactory managementServicesFactory,
+            IMessagesProvider messagesProvider)
             : base(botService, botSettings, callbackDataProvider)
         {
             _messagesHelperService = messagesHelperService;
             _argumentParseHelperService = argumentParseHelperService;
+            _managementServicesFactory = managementServicesFactory;
             _messagesProvider = messagesProvider;
-            _managementService = managementService;
 
-            AddCommandHandler(CallbackDataProvider.GetSetDefaultPhotoCallback<ParentCat>(Enums.PhotosManagementMode.Photos), HandleCallbackAsync<ParentCat>);
-            AddCommandHandler(CallbackDataProvider.GetSetDefaultPhotoCallback<Litter>(Enums.PhotosManagementMode.Photos), HandleCallbackAsync<Litter>);
-            AddCommandHandler(CallbackDataProvider.GetSetDefaultPhotoCallback<Kitten>(Enums.PhotosManagementMode.Photos), HandleCallbackAsync<Kitten>);
+            AddCommandHandler(CallbackDataProvider.GetSetDefaultPhotoCallback<ParentCat>(PhotosType.Photos), HandleCallbackAsync<ParentCat>);
+            AddCommandHandler(CallbackDataProvider.GetSetDefaultPhotoCallback<Litter>(PhotosType.Photos), HandleCallbackAsync<Litter>);
+            AddCommandHandler(CallbackDataProvider.GetSetDefaultPhotoCallback<Kitten>(PhotosType.Photos), HandleCallbackAsync<Kitten>);
         }
 
         private async Task HandleCallbackAsync<TEntity>(CallbackQueryAdapter c, CancellationToken token)
             where TEntity : class, IDisplayableEntity
         {
-            var args = c.CallbackData.Split(CallbackArgsSeparator); //[0]Command, [1]PhotoIndex, [2]Entity Id
+            var args = c.CallbackData.Split(CallbackArgsSeparator); //[0]Command, [1]PhotoId, [2]Entity Id
             var entityId = int.Parse(args.Last());
-            var photoIndex = int.Parse(args[1]);
-            var result = await _managementService.SetDefaultPhotoForEntityAsync<TEntity>(entityId, photoIndex, token);
+            var photoId = int.Parse(args[1]);
+            var managementService = _managementServicesFactory.GetDisplayableEntityManagementService<TEntity>();
+            var result = await managementService.SetDefaultPhotoAsync(entityId, photoId, token);
 
             if (result.Success)
             {
-                var entity = result.Result!;
+                var entity = await managementService.GetEntityAsync(entityId, token);
+
+                if (entity == null)
+                {
+                    await BotService.AnswerCallbackQueryAsync(c.CallbackId, _messagesProvider.CreateEntityNotFoundMessage(), token: token);
+                    return;
+                }
 
                 var key = c.MessageText.Split('\n').Last();
                 var (_, photoMessageIds) = _argumentParseHelperService.ParsePhotosArgs(key);
 
                 await BotService.DeleteMessagesAsync(c.Chat, [.. photoMessageIds, c.MessageId], token);
                 await _messagesHelperService.SendPhotoManagementMessageAsync(c.Chat, entity, token);
-                await BotService.AnswerCallbackQueryAsync(c.CallbackId, _messagesProvider.CreateDefaultPhotoSetForEntityMessage(entity, photoIndex), token: token);
+                await BotService.AnswerCallbackQueryAsync(c.CallbackId, _messagesProvider.CreateDefaultPhotoSetForEntityMessage(entity, photoId), token: token);
             }
             else
             {
